replace type for mysql

+++ SwiftBBS/SwiftBBS Server/BbsCommentRepository.swift
@@ -61,10 +61,10 @@ class BbsCommentRepository : Repository {
     //  row contains b.id, b.bbs_id, b.comment, b.user_id, u.name, b.created_at, b.updated_at
     private func createEntityFromRow(row: Row) -> BbsCommentWithUserEntity {
         return BbsCommentWithUserEntity(
-            id: UInt(row[0] as! UInt64),
-            bbsId: UInt(row[1] as! UInt64),
+            id: UInt(row[0] as! UInt32),
+            bbsId: UInt(row[1] as! UInt32),
             comment: stringFromMySQLText(row[2] as? [UInt8]) ?? "",
-            userId: UInt(row[3] as! UInt64),
+            userId: UInt(row[3] as! UInt32),
             userName: row[4] as! String,
             createdAt: row[5] as! String,
             updatedAt: row[6] as! String
